Refactor:  영역외 클릭시 닫는 동작 useClickOutside 훅으로 분리
